# docs/swagger.yaml
basePath: /
definitions:
  godoc.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  godoc.ErrorsResponse:
    properties:
      messages:
        items:
          type: string
        type: array
    type: object
  request.CreateTodoRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
  request.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  request.RegisterUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  request.UpdateTodoByIdRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
  response.GetTodoByIdResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  response.GetTodoListResponse:
    properties:
      page:
        type: integer
      size:
        type: integer
      todos:
        items:
          $ref: '#/definitions/response.GetTodoByIdResponse'
        type: array
    type: object
  response.GetUserInfoResponse:
    properties:
      email:
        type: string
      id:
        type: integer
    type: object
  response.LoginUserResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  response.UpdateTodoByIdResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Just for practice
  title: Todo Golang Example
  version: "1.0"
paths:
  /todo/{id}:
    delete:
      parameters:
      - description: Request Param
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/godoc.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Xóa todo theo id
      tags:
      - Todo
    get:
      parameters:
      - description: Request Param
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetTodoByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/godoc.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tìm todo bằng id
      tags:
      - Todo
  /todo/create:
    post:
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetTodoByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/godoc.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tạo mới todo
      tags:
      - Todo
  /todo/list:
    get:
      parameters:
      - description: Query Param
        in: query
        name: page
        required: true
        type: integer
      - description: Query Param
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetTodoListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/godoc.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Phân trang danh sách todo
      tags:
      - Todo
  /todo/update/{id}:
    put:
      parameters:
      - description: Request Param
        in: path
        name: id
        required: true
        type: integer
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTodoByIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UpdateTodoByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/godoc.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật todo theo id
      tags:
      - Todo
  /user/info:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetUserInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Thông tin người dùng
      tags:
      - User
  /user/login:
    post:
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoginUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/godoc.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
      summary: Đăng nhập
      tags:
      - User
  /user/register:
    post:
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/godoc.ErrorsResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/godoc.ErrorResponse'
      summary: Đăng ký
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"